requirements.txt
numpy
scipy
joblib
scikit-learn
Pillow
